refactor: club 도메인
