+++ src/docs/asciidoc/exception-api.adoc
@@ -67,7 +67,10 @@
 |잘못된 요청입니다
 
 |`+EVENT_NOT_MANAGED+`
-|폼이 없거나 관리모드를 설정하지 않은 폼입니다.
+|폼이 없거나 관리모드를 설정하지 않은 폼입니다
+
+|`+CAN_NOT_SELF_DEGRADING+`
+|혼자 남은 관리자는 강등될 수 없습니다
 
 
 |====
