+++ src/main/java/com/spaceclub/club/repository/ClubUserRepository.java
@@ -3,7 +3,6 @@ package com.spaceclub.club.repository;
 import com.spaceclub.club.domain.Club;
 import com.spaceclub.club.domain.ClubUser;
 import com.spaceclub.club.domain.ClubUserRole;
-import com.spaceclub.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -13,18 +12,18 @@ public interface ClubUserRepository extends JpaRepository<ClubUser, Long> {
 
     List<ClubUser> findByClub_Id(Long clubId);
 
-    Optional<ClubUser> findByClub_IdAndUser_Id(Long clubId, Long userId);
+    Optional<ClubUser> findByClub_IdAndUserId(Long clubId, Long userId);
 
     int countByClub_IdAndRole(Long clubId, ClubUserRole role);
 
-    List<ClubUser> findByUser_Id(Long userId);
+    List<ClubUser> findByUserId(Long userId);
 
-    boolean existsByClubAndUser(Club club, User user);
+    boolean existsByClubAndUserId(Club club, Long userId);
 
     Long countByClub(Club club);
 
     ClubUser findByClub_IdAndRole(Long clubId, ClubUserRole role);
 
-    Boolean existsByClub_IdAndUser_Id(Long clubId, Long userId);
+    Boolean existsByClub_IdAndUserId(Long clubId, Long userId);
 
 }
