+++ src/main/java/com/spaceclub/global/ExceptionCode.java
@@ -25,7 +25,8 @@ public enum ExceptionCode {
     CLUB_ALREADY_JOINED("이미 해당 클럽에 가입되어 있습니다"),
     ALREADY_BOOKMARKED("이미 북마크한 이벤트입니다"),
     BAD_REQUEST("잘못된 요청입니다"),
-    EVENT_NOT_MANAGED("폼이 없거나 관리모드를 설정하지 않은 폼입니다.");
+    EVENT_NOT_MANAGED("폼이 없거나 관리모드를 설정하지 않은 폼입니다"),
+    CAN_NOT_SELF_DEGRADING("혼자 남은 관리자는 강등될 수 없습니다");
 
     @Getter
     private final String message;
