+++ src/main/java/com/spaceclub/invite/service/InviteCodeGenerator.java
@@ -1,5 +1,6 @@
-package com.spaceclub.invite.domain;
+package com.spaceclub.invite.service;
 
+@FunctionalInterface
 public interface InviteCodeGenerator {
 
     String generateCode();
