+++ src/test/java/com/spaceclub/invite/controller/InviteJoinControllerTest.java
@@ -1,10 +1,10 @@
 package com.spaceclub.invite.controller;
 
 import com.spaceclub.SpaceClubCustomDisplayNameGenerator;
-import com.spaceclub.club.domain.Club;
+import com.spaceclub.club.service.ClubMemberManagerService;
 import com.spaceclub.global.UserArgumentResolver;
 import com.spaceclub.global.interceptor.JwtAuthorizationInterceptor;
-import com.spaceclub.invite.service.InviteService;
+import com.spaceclub.invite.service.InviteJoinService;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +17,8 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
+import java.util.UUID;
+
 import static com.spaceclub.club.ClubTestFixture.club1;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
@@ -24,10 +26,12 @@ import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
+import static org.springframework.restdocs.payload.JsonFieldType.NUMBER;
 import static org.springframework.restdocs.payload.JsonFieldType.STRING;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
@@ -37,7 +41,7 @@ import static org.springframework.security.test.web.servlet.request.SecurityMock
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@WebMvcTest(value = InviteController.class,
+@WebMvcTest(value = InviteJoinController.class,
         excludeFilters = {
                 @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = {
                         JwtAuthorizationInterceptor.class,
@@ -45,47 +49,81 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
         })
 @AutoConfigureRestDocs
 @DisplayNameGeneration(SpaceClubCustomDisplayNameGenerator.class)
-class InviteControllerTest {
+class InviteJoinControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
 
     @MockBean
-    private InviteService inviteService;
+    private InviteJoinService inviteJoinService;
+
+    @MockBean
+    private ClubMemberManagerService clubMemberManagerService;
 
     @MockBean
     private UserArgumentResolver userArgumentResolver;
 
     @Test
     @WithMockUser
-    void 클럽_초대_링크_생성에_성공한다() throws Exception {
+    void 초대_링크를_통해_클럽_가입에_성공한다() throws Exception {
         // given
-        Club club = club1();
-        Long clubId = club.getId();
-
-        given(inviteService.getInviteCode(any(Long.class), any())).willReturn("650d2d91-a8cf-45e7-8a43-a0c798173ecb");
+        String code = UUID.randomUUID().toString();
 
         // when
-        ResultActions actions = mockMvc.perform(post("/api/v1/clubs/{clubId}/invite", clubId)
-                .header(AUTHORIZATION, "token")
-                .with(csrf()));
+        ResultActions actions =
+                mockMvc.perform(post("/api/v1/clubs/invite/{code}", code)
+                        .header(AUTHORIZATION, "token")
+                        .with(csrf()));
 
         // then
         actions.andExpect(status().isOk())
                 .andDo(print())
-                .andDo(document("invite/create",
+                .andDo(document("invite/join",
                         preprocessRequest(prettyPrint()),
                         preprocessResponse(prettyPrint()),
                         pathParameters(
-                                parameterWithName("clubId").description("클럽 ID")
+                                parameterWithName("code").description("클럽 초대 링크 식별자")
                         ),
                         requestHeaders(
                                 headerWithName(AUTHORIZATION).description("유저 액세스 토큰")
                         ),
                         responseFields(
-                                fieldWithPath("inviteLink").type(STRING).description("클럽 초대 링크")
+                                fieldWithPath("clubId").type(NUMBER).description("클럽 ID")
                         )
                 ));
+
+    }
+
+    @Test
+    @WithMockUser
+        // 없음.
+    void 초대_링크를_통해_클럽_가입전_가입_의사를_묻는데_성공한다() throws Exception {
+        // given
+        String code = UUID.randomUUID().toString();
+        given(inviteJoinService.requestToJoinClub(any(String.class))).willReturn(club1());
+
+        // when
+        ResultActions actions =
+                mockMvc.perform(get("/api/v1/clubs/invite/{code}", code));
+
+        // then
+        actions.andExpect(status().isOk())
+                .andDo(print())
+                .andDo(document("invite/requestToJoin",
+                        preprocessRequest(prettyPrint()),
+                        preprocessResponse(prettyPrint()),
+                        pathParameters(
+                                parameterWithName("code").description("클럽 초대 링크 식별자")
+                        ),
+                        responseFields(
+                                fieldWithPath("clubId").type(NUMBER).description("클럽 ID"),
+                                fieldWithPath("name").type(STRING).description("클럽 이름"),
+                                fieldWithPath("info").type(STRING).description("클럽 소개"),
+                                fieldWithPath("memberCount").type(NUMBER).description("클럽 구성원 수"),
+                                fieldWithPath("logoImageUrl").type(STRING).description("클럽 로고 이미지 URL")
+                        )
+                ));
+
     }
 
 }
