+++ src/test/java/com/spaceclub/invite/service/InviteJoinServiceTest.java
@@ -16,7 +16,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Optional;
 
-import static com.spaceclub.club.ClubTestFixture.club1;
 import static com.spaceclub.invite.InviteTestFixture.invite1;
 import static com.spaceclub.user.UserTestFixture.user1;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -26,10 +25,10 @@ import static org.mockito.Mockito.lenient;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayNameGeneration(SpaceClubCustomDisplayNameGenerator.class)
-class InviteServiceTest {
+class InviteJoinServiceTest {
 
     @InjectMocks
-    private InviteService inviteService;
+    private InviteJoinService inviteJoinService;
 
     @Mock
     private InviteRepository inviteRepository;
@@ -49,7 +48,7 @@ class InviteServiceTest {
         given(inviteRepository.findByCode(any(String.class))).willReturn(Optional.of(expiredInvite));
 
         // when, then
-        assertThatThrownBy(() -> inviteService.joinClub("123", 1L))
+        assertThatThrownBy(() -> inviteJoinService.joinClub("123", 1L))
                 .isInstanceOf(IllegalStateException.class);
 
     }
@@ -60,10 +59,10 @@ class InviteServiceTest {
         given(userRepository.findById(any(Long.class))).willReturn(Optional.of(user1()));
         given(inviteRepository.findByCode(any(String.class))).willReturn(Optional.of(invite1()));
         lenient().doReturn(true).when(clubUserRepository)
-                .existsByClubAndUser(any(Club.class), any(User.class));
+                .existsByClubAndUserId(any(Club.class), any(Long.class));
 
         // when, then
-        assertThatThrownBy(() -> inviteService.joinClub("123", 1L))
+        assertThatThrownBy(() -> inviteJoinService.joinClub("123", 1L))
                 .isInstanceOf(IllegalStateException.class);
 
     }
